+++ packages/components/cascader-panel/src/types.ts
@@ -1,13 +1,51 @@
-import type { InjectionKey } from 'vue'
-import type {
-  CascaderConfig,
-  default as CascaderNode,
-  CascaderOption,
-  RenderLabel,
-} from './node'
+import type { InjectionKey, VNode } from 'vue'
+import type { default as CascaderNode } from './node'
 
-export type * from './node'
 export type { CascaderNode }
+export type CascaderNodeValue = string | number
+export type CascaderNodePathValue = CascaderNodeValue[]
+export type CascaderValue =
+  | CascaderNodeValue
+  | CascaderNodePathValue
+  | (CascaderNodeValue | CascaderNodePathValue)[]
+export type CascaderConfig = Required<CascaderProps>
+export type ExpandTrigger = 'click' | 'hover'
+export type isDisabled = (data: CascaderOption, node: CascaderNode) => boolean
+export type isLeaf = (data: CascaderOption, node: CascaderNode) => boolean
+export type Resolve = (dataList?: CascaderOption[]) => void
+export type LazyLoad = (node: CascaderNode, resolve: Resolve) => void
+export interface RenderLabelProps {
+  node: CascaderNode
+  data: CascaderOption
+}
+export type RenderLabel = (props: RenderLabelProps) => VNode | VNode[]
+export interface CascaderOption extends Record<string, unknown> {
+  label?: string
+  value?: CascaderNodeValue
+  children?: CascaderOption[]
+  disabled?: boolean
+  leaf?: boolean
+}
+
+export interface CascaderProps {
+  expandTrigger?: ExpandTrigger
+  multiple?: boolean
+  checkStrictly?: boolean
+  emitPath?: boolean
+  lazy?: boolean
+  lazyLoad?: LazyLoad
+  value?: string
+  label?: string
+  children?: string
+  disabled?: string | isDisabled
+  leaf?: string | isLeaf
+  hoverThreshold?: number
+  checkOnClickNode?: boolean
+  checkOnClickLeaf?: boolean
+  showPrefix?: boolean
+}
+
+export type Nullable<T> = null | T
 
 export interface Tag {
   node?: CascaderNode
